--- tools/c7n_huaweicloud/c7n_huaweicloud/apig_sdk/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ["signer"]
