+++ tools/c7n_huaweicloud/c7n_huaweicloud/utils/signer.py
